allow setting allowed routes on proxy
