+++ litellm/proxy/proxy_config.yaml
@@ -13,3 +13,4 @@ litellm_settings:
 general_settings: 
  master_key: sk-1234 
  custom_auth: example_config_yaml.custom_auth_basic.user_api_key_auth
+ allowed_routes: []
